Moving source code to `io.guedes` package
