+++ app/src/main/java/io/guedes/notes/app/common/ui/activity/ActivityRequestCodes.kt
@@ -1,0 +1,5 @@
+package io.guedes.notes.app.common.ui.activity
+
+object ActivityRequestCodes {
+    const val CREATE_NOTE = 1
+}
