+++ app/src/main/java/io/guedes/notes/app/note/view/CreateNoteView.kt
@@ -1,0 +1,8 @@
+package io.guedes.notes.app.note.view
+
+import io.guedes.notes.domain.model.Note
+
+interface CreateNoteView {
+    fun noteCreated(note: Note)
+    fun invalidForm()
+}
