+++ app/src/main/java/io/guedes/notes/app/note/view/ListNotesView.kt
@@ -1,0 +1,8 @@
+package io.guedes.notes.app.note.view
+
+import io.guedes.notes.domain.model.Note
+
+interface ListNotesView {
+    fun showNote(note: Note)
+    fun openCreateNoteForm()
+}
