+++ app/src/main/java/io/guedes/notes/domain/model/Label.kt
@@ -1,0 +1,10 @@
+package io.guedes.notes.domain.model
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Label(
+    val id: Long,
+    val title: String
+) : Parcelable
