+++ app/src/main/java/io/guedes/notes/domain/model/Note.kt
@@ -1,0 +1,15 @@
+package io.guedes.notes.domain.model
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+import java.util.*
+
+@Parcelize
+data class Note(
+    val title: String,
+    val content: String? = null,
+    val label: Label? = null,
+    val remindOn: Date? = null
+) : Parcelable {
+    private constructor() : this("", "") // For Firebase
+}
