--- app/src/main/kotlin/com/fgsguedes/notes/app/common/ui/activity/ActivityRequestCodes.kt
@@ -1,5 +1,0 @@
-package com.fgsguedes.notes.app.common.ui.activity
-
-object ActivityRequestCodes {
-    const val CREATE_NOTE = 1
-}
