--- app/src/main/kotlin/com/fgsguedes/notes/app/note/view/CreateNoteView.kt
@@ -1,8 +1,0 @@
-package com.fgsguedes.notes.app.note.view
-
-import com.fgsguedes.notes.domain.model.Note
-
-interface CreateNoteView {
-    fun noteCreated(note: Note)
-    fun invalidForm()
-}
