--- app/src/main/kotlin/com/fgsguedes/notes/app/note/view/ListNotesView.kt
@@ -1,8 +1,0 @@
-package com.fgsguedes.notes.app.note.view
-
-import com.fgsguedes.notes.domain.model.Note
-
-interface ListNotesView {
-    fun showNote(note: Note)
-    fun openCreateNoteForm()
-}
