--- app/src/main/kotlin/com/fgsguedes/notes/domain/model/Label.kt
@@ -1,10 +1,0 @@
-package com.fgsguedes.notes.domain.model
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Label(
-    val id: Long,
-    val title: String
-) : Parcelable
